+++ changelog.md
@@ -13,6 +13,7 @@ Current version: 0.7
     * Add language development measures: `MLUm()`, `MLUw()`, `TTR()`, `IPSyn()`
     * Add `concordances()` for search queries and concordances
     * Allow regular expression matching when specifying participants in all relevant `SingleReader` and `Reader` methods
+    * Add the `month` parameter in `age()`
     * Add output formats for dependency graphs: `to_tikz()` and `to_conll()`
     * Distinguish `participant_name` and `participant_role` in metadata
     * The `@Languages` header contents are treated as a list but now a set now for ordering wrt language dominance